new file:   .gitignore 	modified:   .vscode/settings.json 	modified:   include/HttpRequest.h 	renamed:    README.md -> old/README.md 	new file:   old/httpdocs/post.cgi 	new file:   old/httpdocs/post.html 	new file:   old/httpdocs/test.html 	renamed:    image/1-1.png -> old/image/1-1.png 	renamed:    image/1.png -> old/image/1.png 	renamed:    image/2.png -> old/image/2.png 	renamed:    image/3.png -> old/image/3.png 	renamed:    image/4.png -> old/image/4.png 	renamed:    image/5.png -> old/image/5.png 	renamed:    image/myhttp.png -> old/image/myhttp.png 	modified:   src/HttpRequest.cpp 	modified:   src/HttpServer.cpp

+++ include/HttpRequest.h
@@ -34,7 +34,7 @@ class HttpRequest
     static constexpr int MAX_LINE_LENGTH = 1024; // 定义最大行长度常量
 
     // 辅助函数
-    static int getLine(int sock, std::string &buf);
+    static size_t getLine(int sock, std::string &buf);
 
     // 检查文件访问权限
     bool checkFileAccess();
